refactor(transformers): consume new opaque features + Tier 3 mechanical wins
Six independent low-risk simplifications grouped into one pass:

* **prefix_accountant.process** — drop the underscore now that
  ``Accountant.process`` is a public attribute (opaque#238).  Single
  trainer line; no behavior change.
* **label_smoothing as loss kwarg** — push
  ``args.label_smoothing_factor`` into ``inputs`` so the now-merged
  ``Opaque_CrossEntropyLoss`` smoothing support (opaque#239) and the
  ``fused_linear_cross_entropy`` kernel apply smoothing natively.  The
  trainer-side rebuild on top of exposed logits stays as the fallback
  for the no-patches case (HF's native CE silently drops the kwarg);
  it's idempotent w.r.t. the kernel path.  Drop the now-vacuous
  ``_smoothing_kernel_warned`` field + in-vmap warning (the warning
  fired when the fused-linear path couldn't honor smoothing, which the
  kernel now does directly).  Drops the dead ``import warnings``.
* **``_is_retryable_oom`` → ``torch.OutOfMemoryError`` isinstance**
  (torch >= 2.4) — replaces three-spelling substring match that was
  fragile across torch versions.  Three test OOM-simulation sites
  updated to raise the typed exception.
* **``_maybe_prime_collate`` exception narrowing** — keep priming
  behavior (collator side-effect caches a template row for empty
  Poisson batches), narrow ``except Exception`` to the actual failure
  modes (``TypeError`` for streaming datasets, ``(TypeError,
  IndexError, KeyError)`` for row lookup, ``(TypeError, ValueError,
  KeyError)`` for collator rejection).  Unrelated exceptions now
  surface instead of being swallowed.
* **``torch.compile fullgraph`` lazy fallback** — try
  ``fullgraph=True`` first, lazily downgrade to ``fullgraph=False``
  with a warning if the closure can't be traced fully.
  ``fullgraph=True`` first surfaces graph-break regressions the user
  explicitly opted into compiling against, instead of silently
  fragmenting and falling back to eager.
* **``DPTrainerState`` step-field types unified to ``int``** —
  ``logging_steps`` / ``eval_steps`` / ``save_steps`` were declared as
  ``float`` / ``float | None`` / ``int`` respectively but all carry the
  same resolved-cadence semantics post-``compute_steps``.  Declare all
  three as ``int`` and ``int(num_steps)`` inside ``compute_steps`` so
  ``state`` always sees integer counts regardless of how the user
  supplied them on ``args``.

+++ packages/opaque-transformers/src/opaque/api/transformers/trainer/_state.py
@@ -42,8 +42,12 @@ class DPTrainerState:
     epoch: float = 0.0
     global_step: int = 0
     max_steps: int = 0
-    logging_steps: float = 500
-    eval_steps: float | None = 500
+    # Resolved cadence in absolute steps.  ``compute_steps`` pre-resolves
+    # fractional values (e.g. ``logging_steps=0.5``) and ``None`` to ints
+    # before storing here; the on-state representation is uniformly
+    # ``int`` for all three.
+    logging_steps: int = 500
+    eval_steps: int | None = 500
     save_steps: int = 500
     train_batch_size: int | None = None
     num_train_epochs: int = 0
@@ -80,7 +84,10 @@ class DPTrainerState:
                 continue
             if num_steps < 1:
                 num_steps = math.ceil(self.max_steps * num_steps)
-            setattr(self, f"{kind}_steps", num_steps)
+            # Coerce to ``int`` so the resolved cadence on state is
+            # uniformly typed (the dataclass declares int fields; user
+            # ``args`` may carry the value as float).
+            setattr(self, f"{kind}_steps", int(num_steps))
 
     def to_json(self) -> dict[str, Any]:
         """Return a JSON-compatible dict for ``trainer_state.json``."""
